Make common SearchBars into a common component
